+++ projects/fhi-angular-components/src/lib/fhi-modal/fhi-modal-action-button.model.ts
@@ -1,0 +1,5 @@
+export interface FhiModalActionButton {
+  name: string;
+  disabled?: boolean;
+  primary?: boolean;
+}